backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# OpenAI SDK
openai==1.3.5

# Environment variables
python-dotenv==1.0.0

# Image processing
Pillow==10.1.0

# Validation and serialization
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP client
httpx==0.25.2

# Utilities
python-dateutil==2.8.2

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9  # PostgreSQL adapter

# Testing (optional)
# pytest==7.4.3
# pytest-asyncio==0.21.1
# pytest-cov==4.1.0

# Logging
structlog==23.2.0

# Google Play API (optional, for Android receipt verification)
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
